tortoise_orm/pyproject.toml
[tool.poetry]
name = "tortoise-orm-test"
version = "0.1.0"
description = ""
authors = ["Dominux <kir.gimranov@gmail.com>"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.8"
tortoise-orm = {extras = ["asyncpg"], version = "^0.19.2"}
fastapi = {extras = ["all"], version = "^0.85.0"}


[tool.poetry.group.dev.dependencies]
pytest = "^7.1.3"
pytest-asyncio = "^0.19.0"
httpx = "^0.23.0"
black = "^22.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
